backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.2
supabase==2.4.6
openai==1.30.1
python-dotenv==1.0.1